Restricted module-info to hide internals
Exposing just the UI & Tools packages only.  Note that this includes
all classes in those packages (but not subpackages), so we might want
to re-org in the future to have only the Main classes public.
The bulk of the args needed to expose our internals to jUnit have been
moved from the Gradle build file to a test_suite.args file.